+++ huey_consumer.py
@@ -11,5 +11,6 @@ from src.core.queue_config import huey
 # To make tasks discoverable, we need to import the modules where they are defined.
 import workers.transcription_worker
 import workers.logging_worker
+import workers.hardware_monitor_worker
 
 # The `huey` object itself is the primary export for the consumer.